Change to 5-based columns
This is required in Sphinx 1.4.4.
I did not track down what that last element is meant to be used for,
but several other projects (and sphinx itself when it encounters
the old 4 column form) set the fifth element to None.
Setting it to a string does not cause that string to be part of the
docs anywhere, so None is good enough for now.